chore: extract service-worker to its own package
