+++ packages/embeddable/package.json
@@ -23,7 +23,6 @@
     "build": "run-s clean build:*",
     "build:dev": "vite build",
     "build:prod": "vite build --minify",
-    "build:sw": "cross-env ENTRY=sw vite build --minify",
     "start": "yarn build:dev --watch",
     "cypress": "cypress run",
     "cypress:headed": "cypress open",
@@ -36,7 +35,6 @@
     "@floating-ui/react": "^0.26.24",
     "@magicbell/magicbell-react": "11.0.2",
     "error-stack-parser": "^2.1.4",
-    "idb": "^8.0.0",
     "preact": "^10.24.0",
     "react-frame-component": "^5.2.6",
     "react-use": "^17.5.0",
