+++ packages/service-worker/README.md
@@ -1,0 +1,5 @@
+**WIP**
+
+This is the MagicBell service worker, as hosted on https://assets.magicbell.io/web-push-notifications/sw.js.
+
+Despite it being the same, we currently don't deploy it from this package just yet.
