+++ packages/service-worker/tsconfig.build.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "noEmit": true
+  },
+  "include": ["./src/**/*.ts", "./src/**/*.mts", "./src/**/*.tsx", "./src/**/*.json"]
+}
