+++ tsconfig.json
@@ -26,6 +26,7 @@
       "@magicbell/magicbell-react": ["packages/react/src"],
       "@magicbell/project-client": ["packages/project-client/src"],
       "@magicbell/react-headless": ["packages/react-headless/src"],
+      "@magicbell/service-worker": ["packages/service-worker/src"],
       "@magicbell/user-client": ["packages/user-client/src"],
       "@magicbell/utils": ["packages/utils/src"],
       "@magicbell/webpush": ["packages/webpush/src"],
